config: define ACL size as a signed int to avoid warnings

+++ config.h
@@ -59,7 +59,7 @@
 #define HTTP_RESPONSE_MAX 2048
 
 // Max number of IPs in the ACL. At HTTP_RESPONSE_MAX=512 should give us 31
-#define ACL_IP_MAX ((unsigned int) ((HTTP_RESPONSE_MAX - 19 + 16) / 16))
+#define ACL_IP_MAX ((HTTP_RESPONSE_MAX - 19 + 16) / 16)
 
 struct http_response {
   int value; // return value. Zero on success, otherwise respects errno.h
